[GitHub Actions] Code Coverage Output (#4)
* code coverage output

* code coverage output

* code coverage output

* code coverage output

* code coverage output

* code coverage output

* code coverage output

* tweaked github actions

* added test cases to increase coverage

* added assertion to test case

* tweaked test case

* tweaked test case

+++ src/HttpBgCallback.php
@@ -62,8 +62,8 @@ class HttpBgCallback extends Command
     {
         $connectionTimeout   = intval($this->argument('connection_timeout'));
         $totalRequestTimeout = intval($this->argument('total_request_timeout'));
-        $curlTimeConnect     = intval($this->argument('curl_time_connect'));
-        $curlTimeTotal       = intval($this->argument('curl_time_total'));
+        $curlTimeConnect     = floatval($this->argument('curl_time_connect'));
+        $curlTimeTotal       = floatval($this->argument('curl_time_total'));
 
         if ($curlErrorCode === 28) {
             return true;
